requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-extensions==3.1.5
django-queryset-csv==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-extra-fields==3.2.1
idna==3.3
jdatetime==3.8.2
psycopg2-binary==2.9.2
pyaes==1.6.1
pyasn1==0.4.8
pyTelegramBotAPI==4.3.1
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
tornado==6.1
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.7
